gateway: Split 'http' and 'opaq' modules
To setup for upcoming stack refactoring, this change organizes the
gateway crate into `http` and `opaq` modules.
